requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
frozenlist==1.7.0
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
multidict==6.6.3
packaging==25.0
pluggy==1.6.0
propcache==0.3.2
psycopg2-binary==2.9.10
Pygments==2.19.2
pyTelegramBotAPI==4.28.0
pytest==8.4.1
python-dotenv==1.1.1
requests==2.32.4
telebot==0.0.5
typing_extensions==4.14.1
urllib3==2.5.0
wheel==0.45.1
yarl==1.20.1
